rename: GameResult, GameResultGenerator를 result 패키지로 분리 및 그에 따른 변경 사항
