+++ src/main/java/racingcar/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.result;
 
 public class GameResult {
